requirements.txt
couchbase==4.3.5
python-dotenv==1.1.0
sentence-transformers==3.4.1
langchain-huggingface==0.1.2